chore: move helpers to modules
