+++ src/modules/getRandom.ts
