+++ src/modules/sendLove.ts
@@ -3,10 +3,10 @@ import { EmbedBuilder, Message } from "discord.js";
 import { hearts } from "../data/hearts";
 import { loveData } from "../data/loveData";
 import { phrases } from "../data/phrases";
-import { getRandom } from "../helpers/getRandom";
 import { Bot } from "../interfaces/Bot";
 import { errorHandler } from "../utils/errorHandler";
 
+import { getRandom } from "./getRandom";
 import { validateLove } from "./validateLove";
 
 /**
